完善隔离见证

+++ pop-coin-system/src/main/java/com/pop/popcoinsystem/data/transaction/constant.java
@@ -4,7 +4,7 @@ public class constant {
 
 
     // 定义支持的交易版本
-    public static final long VERSION_1 = 1L;  // 基础交易版本
-    public static final long VERSION_2 = 2L;  // 支持时间锁等高级功能
+    public static final int VERSION_1 = 1;  // 基础交易版本
+    public static final int VERSION_2 = 2;  // 支持时间锁等高级功能
 
 }
